create and test getComputerChoice function
The function returns "rock", "paper" or "scissors"
randomly. It uses the Math.random() function,
which is not perfectly random, but good enough for
this game.

The function was tested by logging its result to
the console 100 times and ensuring all results are
within the correct range. It should be impossible
for it to return a result with an index >2.